fixed styling for tab character
this may seem to be splitting hairs, but `tab` means typing the characters `t` and `a` and `b`, which is obviously not what is meant.  I'm not sure what to do with the tab inside the code block though, as there's no way (that I know of) to render HTML inside a code block, at least not the way Github implements them.